separate handler functions from routes
